Images in styled text overflowing text above it when converted to PDF using HTMLWorker
When a paragraph containing an image is converted to PDF using HTMLWorker height of the line is calculated only by the size of the text in it.

<p>Here is an image from the beautiful Istanbul bosphorous :) <img src="http://www.romancehotel.com/Admin/Images/TouristicPlaces/b80c1486-8ec5-4b56-812a-bc1681701539.jpg"> </p>

Luckily, iText already provides a way to trigger the calculation of the line considering the image height as well with a flag as shown below

new Chunk(img, 0, 0, true);

I have updated the HTMLWorker to use the flag since it seems to be the default behavior for HTML as well. If you run the above HTML source in a browser you can verify that line height is of the height of the image which is taller.